refactor: separate generateRpcSpec into its own util file
