+++ scripts/build.ts
@@ -1,92 +1,19 @@
-import type {
-  JSONSchema,
-  Methods,
-  OpenrpcDocument,
-  SchemaComponents,
-} from "@open-rpc/meta-schema";
-import { dereferenceDocument } from "@open-rpc/schema-utils-js";
-import { cp, mkdirSync, readFileSync, readdirSync, writeFileSync } from "fs";
-import yaml from "js-yaml";
-import mergeAllOf from "json-schema-merge-allof";
-import { DerefedOpenRpcDoc } from "../src/types/OpenRpc";
+import { cp, mkdirSync, readdirSync } from "fs";
+import { generateRpcSpec } from "../src/utils/generateRpcSpec";
 
-const allChainsBase = "src/openrpc-schemas/chains";
-const allChainFiles = readdirSync(allChainsBase);
+const allChainsDir = "src/openrpc-schemas/chains";
+const allChainFiles = readdirSync(allChainsDir);
 
 const outputDir = "docs/api-specs/chains";
 
 mkdirSync(outputDir, { recursive: true });
 
-const generateChainSpec = async (chainFile: string) => {
-  const chainBase = `${allChainsBase}/${chainFile}`;
-  const componentsBase = `${chainBase}/components`;
-  const methodsBase = `${chainBase}/methods`;
-
-  const componentsFiles = readdirSync(componentsBase);
-  const methodsFiles = readdirSync(methodsBase);
-
-  const unsortedMethods = methodsFiles.reduce<Methods>((acc, file) => {
-    const raw = readFileSync(`${methodsBase}/${file}`).toString();
-    const parsed = yaml.load(raw) as Methods;
-
-    return [...acc, ...parsed];
-  }, []);
-
-  const methods = unsortedMethods.sort((a, b) => {
-    if ("name" in a && "name" in b) {
-      return a.name.localeCompare(b.name);
-    }
-    return 0;
-  });
-
-  const schemas = componentsFiles.reduce<SchemaComponents>((acc, file) => {
-    const raw = readFileSync(`${componentsBase}/${file}`).toString();
-    const parsed = yaml.load(raw) as SchemaComponents;
-
-    return { ...acc, ...parsed };
-  }, {});
-
-  const baseRaw = readFileSync(`${chainBase}/base.yaml`).toString();
-  const base = yaml.load(baseRaw) as Pick<
-    OpenrpcDocument,
-    "info" | "externalDocs" | "servers"
-  >;
-
-  const doc: OpenrpcDocument = {
-    openrpc: "1.2.4",
-    ...base,
-    methods,
-    components: {
-      schemas,
-    },
-  };
-
-  const spec = (await dereferenceDocument(doc)) as DerefedOpenRpcDoc;
-
-  delete spec.components; // once dereferenced, components are no longer needed
-
-  spec.methods.forEach((method) => {
-    method.params.forEach((param) => {
-      if (typeof param.schema !== "boolean") {
-        param.schema = mergeAllOf(param.schema) as JSONSchema;
-      }
-    });
-
-    if (method.result && typeof method.result.schema !== "boolean") {
-      method.result.schema = mergeAllOf(method.result.schema) as JSONSchema;
-    }
-  });
-
-  writeFileSync(
-    `${outputDir}/${chainFile}.json`,
-    JSON.stringify(spec, null, "\t")
-  );
-};
+allChainFiles.forEach((chain) =>
+  generateRpcSpec(allChainsDir, outputDir, chain)
+);
 
 cp("src/markdown", "docs/markdown", { recursive: true }, (e) => {
   if (e) {
     console.error(e);
   }
 });
-
-allChainFiles.forEach(generateChainSpec);
